+++ cumulusci/cli/cci.py
@@ -4,19 +4,16 @@ import code
 import json
 import re
 import os
-import platform
 import pdb
 import sys
 import time
 import traceback
 import runpy
-import webbrowser
 import contextlib
 from datetime import datetime
 from pathlib import Path
 
 import click
-import github3
 import pkg_resources
 import requests
 from rst2ansi import rst2ansi
@@ -25,9 +22,7 @@ import cumulusci
 from cumulusci.core.config import ServiceConfig
 from cumulusci.core.config import TaskConfig
 from cumulusci.core.config import UniversalConfig
-from cumulusci.core.github import create_gist, get_github_api
 from cumulusci.core.exceptions import (
-    CumulusCIException,
     CumulusCIUsageError,
     ServiceNotConfigured,
     FlowNotFoundError,
@@ -46,8 +41,9 @@ from cumulusci.utils import get_cci_upgrade_command
 from cumulusci.utils.logging import tee_stdout_stderr
 from cumulusci.utils.yaml.cumulusci_yml import cci_safe_load
 
-
 from .logger import init_logger, get_tempfile_logger
+
+from .error import error
 from .org import org
 from .project import project
 from .runtime import pass_runtime
@@ -274,22 +270,9 @@ def shell(runtime, script=None, python=None):
         code.interact(local=variables)
 
 
-GIST_404_ERR_MSG = """A 404 error code was returned when trying to create your gist.
-Please ensure that your GitHub personal access token has the 'Create gists' scope."""
-
-
-def get_context_info():
-    host_info = platform.uname()
-
-    info = []
-    info.append(f"CumulusCI version: {cumulusci.__version__}")
-    info.append(f"Python version: {sys.version.split()[0]} ({sys.executable})")
-    info.append(f"Environment Info: {host_info.system} / {host_info.machine}")
-    return "\n".join(info)
-
-
 # Top Level Groups
 
+cli.add_command(error)
 cli.add_command(project)
 cli.add_command(org)
 
@@ -307,29 +290,6 @@ def flow():
 @cli.group("service", help="Commands for connecting services to the keychain")
 def service():
     pass
-
-
-@cli.group("error", short_help="Get or share information about an error")
-def error():
-    """
-    Get or share information about an error
-
-    If you'd like to dig into an error more yourself,
-    you can get the last few lines of context about it
-    from `cci error info`.
-
-    If you'd like to submit it to a developer for conversation,
-    you can use the `cci error gist` command. Just make sure
-    that your GitHub access token has the 'create gist' scope.
-
-    If you'd like to regularly see stack traces, set the `show_stacktraces`
-    option to `True` in the "cli" section of `~/.cumulusci/cumulusci.yml`, or to
-    see a stack-trace (and other debugging information) just once, use the `--debug`
-    command line option.
-
-    For more information on working with errors in CumulusCI visit:
-    https://cumulusci.readthedocs.io/en/latest/features.html#working-with-errors
-    """
 
 
 # Commands for group: service
@@ -878,81 +838,6 @@ def flow_run(runtime, flow_name, org, delete_org, debug, o, skip, no_prompt):
                 "Scratch org deletion failed.  Ignoring the error below to complete the flow:"
             )
             click.echo(str(e))
-
-
-CCI_LOGFILE_PATH = Path.home() / ".cumulusci" / "logs" / "cci.log"
-
-
-@error.command(
-    name="info",
-    help="Outputs the most recent traceback (if one exists in the most recent log)",
-)
-@click.option("--max-lines", "-m", type=int)
-def error_info(max_lines: int = 0):
-    if not CCI_LOGFILE_PATH.is_file():
-        click.echo(f"No logfile found at: {CCI_LOGFILE_PATH}")
-    else:
-        output = lines_from_traceback(
-            CCI_LOGFILE_PATH.read_text(encoding="utf-8"), max_lines
-        )
-        click.echo(output)
-
-
-def lines_from_traceback(log_content: str, max_lines: int = 0) -> str:
-    """Returns the the last max_lines of the logfile,
-    or the whole traceback, whichever is shorter. If
-    no stacktrace is found in the logfile, the user is
-    notified.
-    """
-    stacktrace_start = "Traceback (most recent call last):"
-    if stacktrace_start not in log_content:
-        return f"\nNo stacktrace found in: {CCI_LOGFILE_PATH}\n"
-
-    stacktrace = ""
-    for i, line in enumerate(reversed(log_content.split("\n")), 1):
-        stacktrace = "\n" + line + stacktrace
-        if stacktrace_start in line:
-            break
-        if i == max_lines:
-            break
-
-    return stacktrace
-
-
-@error.command(name="gist", help="Creates a GitHub gist from the latest logfile")
-@pass_runtime(require_project=False, require_keychain=True)
-def gist(runtime):
-    if CCI_LOGFILE_PATH.is_file():
-        log_content = CCI_LOGFILE_PATH.read_text(encoding="utf-8")
-    else:
-        log_not_found_msg = """No logfile to open at path: {}
-        Please ensure you're running this command from the same directory you were experiencing an issue."""
-        error_msg = log_not_found_msg.format(CCI_LOGFILE_PATH)
-        click.echo(error_msg)
-        raise CumulusCIException(error_msg)
-
-    last_cmd_header = "\n\n\nLast Command Run\n================================\n"
-    filename = f"cci_output_{datetime.utcnow()}.txt"
-    files = {
-        filename: {"content": f"{get_context_info()}{last_cmd_header}{log_content}"}
-    }
-
-    try:
-        gh = runtime.keychain.get_service("github")
-        gist = create_gist(
-            get_github_api(gh.username, gh.password or gh.token),
-            "CumulusCI Error Output",
-            files,
-        )
-    except github3.exceptions.NotFoundError:
-        raise CumulusCIException(GIST_404_ERR_MSG)
-    except Exception as e:
-        raise CumulusCIException(
-            f"An error occurred attempting to create your gist:\n{e}"
-        )
-    else:
-        click.echo(f"Gist created: {gist.html_url}")
-        webbrowser.open(gist.html_url)
 
 
 def normalize_option_name(k):
